refactor(data): extract BaseReportTemplateQuerySet to reduce duplication
Create base QuerySet for report templates with shared filtering methods.
Simplify SpecialtyReportTemplate manager to use standard Django pattern.
